web-auto/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
python-multipart>=0.0.9,<1.0
requests>=2.31,<3.0
pillow>=10.0,<12.0
opencv-python-headless>=4.8,<5.0
